Modified the navigation bar to include sign-in/sign-out buttons
The state of the navigavtion bar depends on the user's logged in
status to make sure the relevant buttons are visible and have the
proper links. If the user is logged in, show the 'profile' and
'sign out' buttons, otherwise show the 'sign in' button.

Changed some styling for the navigation bar to accommodate the new
buttons.